waste-classification-main/requirements.txt
flask>=2.0.0
tensorflow>=2.10.0
pillow>=9.0.0
numpy>=1.21.0

